Add '.0' to all the '0' and '1' in the expected-output files
I think that since Lua5.3, floating-point round numbers (such as 0.0
or 1.0) are correctly converted to strings with the trailing ".0"
(such as "0.0" and "1.0"), as opposed to Lua5.2 (which only prints
the integer part - "0" or "1").

Therefore the generated code changes with Lua>5.2, and the existing
tests fail.

We know that these tests based on verbatim comparisons are not
flexible, however expecting "1.0" is probably more sensible than
"1", even when the actual constant is a round 1. That is, this
commit is still a slight logical improvement, besides the fact
that it makes the tests to pass.

A more robust approach would be to have explicit control
on how a number is converted to text, instead of relying on the
internal behaviour of the language.